test(boot): break out `vtransfer.test.ts`
